+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 - Add link at pub.dev back to the GitHub repository.
+- Widget builder: do not pop more than necessary routes when the underlying
+  stream is closed.
 
 ## 0.4.0